+++ src/main/java/gift/dto/category/request/UpdateCategoryRequest.java
@@ -1,4 +1,4 @@
-package gift.dto.category;
+package gift.dto.category.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
