+++ src/main/java/gift/dto/product/request/UpdateProductRequest.java
@@ -1,4 +1,4 @@
-package gift.dto.product;
+package gift.dto.product.request;
 
 import gift.validation.annotation.RestrictedKeyword;
 import jakarta.validation.constraints.NotBlank;
