+++ src/main/java/gift/dto/product/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package gift.dto.product;
+package gift.dto.product.response;
 
 public record ProductResponse(
     Long id,
