+++ src/main/java/gift/dto/user/response/UserResponse.java
@@ -1,4 +1,4 @@
-package gift.dto.user;
+package gift.dto.user.response;
 
 public record UserResponse(
     String token
