+++ src/main/java/gift/dto/wish/request/AddWishRequest.java
@@ -1,10 +1,9 @@
-package gift.dto.wish;
+package gift.dto.wish.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-public record UpdateWishRequest(
-    @NotNull Long id,
+public record AddWishRequest(
     @NotNull Long productId,
     @Positive(message = "상품 수량은 0보다 큰 수이어야 합니다.") Integer quantity
 ) {
