+++ src/main/java/gift/service/UserService.java
@@ -1,8 +1,8 @@
 package gift.service;
 
-import gift.dto.user.UserLoginRequest;
-import gift.dto.user.UserRegisterRequest;
-import gift.dto.user.UserResponse;
+import gift.dto.user.request.UserLoginRequest;
+import gift.dto.user.request.UserRegisterRequest;
+import gift.dto.user.response.UserResponse;
 import gift.entity.User;
 import gift.exception.InvalidTokenException;
 import gift.exception.user.UserAlreadyExistException;
