+++ src/main/java/gift/util/mapper/WishMapper.java
@@ -1,6 +1,6 @@
 package gift.util.mapper;
 
-import gift.dto.wish.WishResponse;
+import gift.dto.wish.response.WishResponse;
 import gift.entity.Product;
 import gift.entity.Wish;
