+++ src/test/java/gift/unit/service/ProductServiceTest.java
@@ -9,9 +9,9 @@ import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import gift.dto.product.CreateProductRequest;
-import gift.dto.product.ProductResponse;
-import gift.dto.product.UpdateProductRequest;
+import gift.dto.product.request.CreateProductRequest;
+import gift.dto.product.response.ProductResponse;
+import gift.dto.product.request.UpdateProductRequest;
 import gift.entity.Product;
 import gift.exception.product.ProductNotFoundException;
 import gift.repository.ProductRepository;
